+++ src/main/java/uk/gov/ea/datareturns/domain/jpa/dao/UniqueIdentifierAliasDao.java
@@ -3,13 +3,7 @@ package uk.gov.ea.datareturns.domain.jpa.dao;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Repository;
-import uk.gov.ea.datareturns.domain.jpa.entities.UniqueIdentifier;
 import uk.gov.ea.datareturns.domain.jpa.entities.UniqueIdentifierAlias;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * DAO for return periods
@@ -19,8 +13,6 @@ import java.util.stream.Collectors;
 @Repository
 @Scope(ConfigurableBeanFactory.SCOPE_SINGLETON)
 public class UniqueIdentifierAliasDao extends EntityDao<UniqueIdentifierAlias> {
-
-    private volatile Map<String, Set<String>> cacheAliasByBaseName = null;
 
     public UniqueIdentifierAliasDao() {
         super(UniqueIdentifierAlias.class);
@@ -34,32 +26,6 @@ public class UniqueIdentifierAliasDao extends EntityDao<UniqueIdentifierAlias> {
     // Just hit the base cache
     public UniqueIdentifierAlias getByNameRelaxed(String name) {
         return getKeyCache().get(name);
-    }
-
-    private Map<String, Set<String>> getAliasCache() {
-        if (cacheAliasByBaseName == null) {
-            synchronized (this) {
-                if (cacheAliasByBaseName == null) {
-                    LOGGER.info("Build base name cache of: UniqueIdentifierAlias");
-                    List<UniqueIdentifierAlias> list = super.list();
-                    cacheAliasByBaseName = list.stream().collect(
-                            Collectors.groupingBy(e -> e.getUniqueIdentifier().getName(),
-                                    Collectors.mapping(e -> e.getName(), Collectors.toSet())
-                            )
-                    );
-                }
-            }
-        }
-        return cacheAliasByBaseName;
-    }
-
-    /**
-     * Get list of alias names for a given UniqueIdentifier
-     * @param uniqueIdentifier
-     * @return Alias names
-     */
-    public Set<String> getAliasNames(UniqueIdentifier uniqueIdentifier) {
-        return getAliasCache().get(uniqueIdentifier.getName());
     }
 
 }
